reverse: convert DX8TextureCategoryClass::Add_Mesh at 0x009471F0
Identity proven by matched caller Insert_To_Texture_Category at
0x009485C0, which calls this symbol twice (existing symbols.csv pin
at 0x009471F0), plus the ZH dx8renderer.cpp twin
DX8TextureCategoryClass::Add_Mesh with the same two-loop polygon
count and index append structure. Second loop uses an unsigned
src_indices subscript, reproducing retail's Peek_Material address
reassociation. Exact 1107B match.

+++ Code/Libraries/Source/WWVegas/WW3D2/DX8TextureCategoryAddMesh.cpp
@@ -1,6 +1,5 @@
-// ?Add_Mesh@DX8TextureCategoryClass@@QAEIAAVVertex_Split_Table@@IIPAVIndexBufferClass@@I@Z
-// partial score=0.99 date=2026-09-20
 // cl: /DNDEBUG /DWIN32 /D_WINDOWS /MD /EHsc /ICode/Libraries/Source/WWVegas/WWMath /ICode/Libraries/Source/WWVegas/WWLib /ICode/Libraries/Source/WWVegas/WW3D2 /ICode/Libraries/Source/WWVegas/WWSaveLoad /ICode/Libraries/Source/WWVegas/Wwutil /ICode/Libraries/Source/WWVegas/WWDownload /ICode/Libraries/Source/Compression /ICode/Libraries/Source/WWVegas/WWDebug /Ireference/shims/sweep /Ibuild/toolchains/dx81/include
+// ?Add_Mesh@DX8TextureCategoryClass@@QAEIAAVVertex_Split_Table@@IIPAVIndexBufferClass@@I@Z
 /*
 ** Command & Conquer Generals Zero Hour(tm)
 ** Copyright 2025 Electronic Arts Inc.
@@ -172,16 +171,18 @@ unsigned DX8TextureCategoryClass::Add_Mesh(
 				unsigned short idx;
 				if (first_polygon == 0xffff)
 					first_polygon = (unsigned short)i;
-				idx = (unsigned short)(src_indices[i][0] + vertex_offset);
+				// Retail indexes the source triangles with an unsigned subscript;
+				// a signed one reassociates the Peek_Material address above.
+				idx = (unsigned short)(src_indices[unsigned(i)][0] + vertex_offset);
 				last_polygon = (unsigned short)i;
 				vmin = MIN(vmin, idx);
 				vmax = MAX(vmax, idx);
 				*dst_indices++ = idx;
-				idx = (unsigned short)(src_indices[i][1] + vertex_offset);
+				idx = (unsigned short)(src_indices[unsigned(i)][1] + vertex_offset);
 				vmin = MIN(vmin, idx);
 				vmax = MAX(vmax, idx);
 				*dst_indices++ = idx;
-				idx = (unsigned short)(src_indices[i][2] + vertex_offset);
+				idx = (unsigned short)(src_indices[unsigned(i)][2] + vertex_offset);
 				vmin = MIN(vmin, idx);
 				vmax = MAX(vmax, idx);
 				*dst_indices++ = idx;
